transfers write to DB and adjust account balances accordingly

+++ tenmo-server/src/main/java/com/techelevator/tenmo/dao/AccountSqlDAO.java
@@ -22,7 +22,7 @@ public class AccountSqlDAO implements AccountDAO {
 	}
 
 	@Override
-	public Double updateBalance(int newBalance, int id) {
+	public Double updateBalance(Double newBalance, int id) {
 		String sql = "UPDATE accounts SET balance = ? WHERE user_id = ?";
 		jdbcTemplate.update(sql, newBalance, id);
